+++ Cargo.toml
@@ -33,6 +33,7 @@ lazy-regex = "2.2.2"
 lazy_static = "1.4.0"
 log = "0.4.6"
 mime = "0.3.0"
+pulldown-cmark = "0.9.0"
 reqwest = { version = "0.11.0", features = ["json"] }
 rust-embed = "*"
 rust_icu_ucol = "1.0.3"
@@ -43,8 +44,3 @@ structopt = "0.3.23"
 syntect = { version = "4.2", default-features = false, features = ["default-fancy"]}
 tokio = { version = "1.0", features = ["macros", "rt-multi-thread"] }
 warp = "0.3.2"
-
-[dependencies.pulldown-cmark]
-default-features = false
-features = ["simd"]
-git = "https://github.com/raphlinus/pulldown-cmark"
