+++ src/server/comment.rs
@@ -119,69 +119,3 @@ impl CommentForm {
         safe_md2html(&self.comment)
     }
 }
-
-#[cfg(test)]
-mod test {
-    #[test]
-    fn markdown_no_html() {
-        let form = super::CommentForm {
-            post: 185,
-            comment: "Hej!\
-                      \r\n\r\nHär är <em>en</em> _kommentar_.\
-                      \r\n\r\n<script>evil</script>"
-                .into(),
-            name: "Rasmus".into(),
-            email: "rasmus@krats.se".into(),
-            url: None,
-        };
-        assert_eq!(
-            &form.html(),
-            "<p>Hej!</p>\
-             \n<p>Här är &lt;em&gt;en&lt;/em&gt; <em>kommentar</em>.</p>\
-             \n&lt;script&gt;evil&lt;/script&gt;",
-        );
-    }
-
-    #[test]
-    fn heading_level() {
-        let form = super::CommentForm {
-            post: 185,
-            comment: "# Rubrik\
-                      \r\n\r\nRubriken ska hamna på rätt nivå.\
-                      \r\n\r\n## Underrubrik
-                      \r\n\r\nOch underrubriken på nivån under."
-                .into(),
-            name: "Rasmus".into(),
-            email: "rasmus@krats.se".into(),
-            url: None,
-        };
-        assert_eq!(
-            &form.html(),
-            "<h4>Rubrik</h4>\
-             \n<p>Rubriken ska hamna på rätt nivå.</p>\
-             \n<h5>Underrubrik</h5>\
-             \n<p>Och underrubriken på nivån under.</p>\n",
-        );
-    }
-    #[test]
-    fn heading_level_2() {
-        let form = super::CommentForm {
-            post: 185,
-            comment: "### Rubrik\
-                      \r\n\r\nRubriken ska hamna på rätt nivå.\
-                      \r\n\r\n#### Underrubrik
-                      \r\n\r\nOch underrubriken på nivån under."
-                .into(),
-            name: "Rasmus".into(),
-            email: "rasmus@krats.se".into(),
-            url: None,
-        };
-        assert_eq!(
-            &form.html(),
-            "<h4>Rubrik</h4>\
-             \n<p>Rubriken ska hamna på rätt nivå.</p>\
-             \n<h5>Underrubrik</h5>\
-             \n<p>Och underrubriken på nivån under.</p>\n",
-        );
-    }
-}
